adk-datareporter/requirements.txt
google-adk==1.2.1
google-cloud-bigquery
gcloud auth application-default login
